+++ src/lib/utility/utilityString.h
@@ -45,7 +45,7 @@ namespace utility
 
 	std::string replace(std::string str, const std::string& from, const std::string& to);
 
-	std::string insertLineBreaksAtBlankSpaces(const std::string& s, int maxLineLength);
+	std::string insertLineBreaksAtBlankSpaces(const std::string& s, size_t maxLineLength);
 
 	std::string trim(const std::string &str);
